\name{Exam}
\alias{Exam}
\docType{data}
\title{
Relation between exam scores and study hours, anxiety scores and average point scores.
}
\description{
The data provide information about students' exam scores, average point score,
the amount of study hours and anxiety scores. 
}

\usage{data(Exam)}
\format{
  A data frame of 20 observations of 4 variables.
  \describe{
    \item{\code{Scores}}{exam scores}
    \item{\code{Hours}}{study hours}
    \item{\code{Anxiety}}{anxiety scores}
    \item{\code{APS}}{average point score}
  }
}
\references{The original source of these data is http://staff.bath.ac.uk/pssiw/stats2/examrevision.sav.
}
\examples{
head(Exam)
}
