\docType{data}
\name{myGORICs}
\alias{myGORICs}
\title{An example of IC values}
\format{
  An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 3 columns.
}
\usage{
  data(myGORICs)
}
\description{
  An example of IC: A matrix with information criteria (AIC, ORIC, GORIC, or GORICA) values of size 4 x 3 (in general: S x 'NrHypos+1', where 'NrHypos+1' stands for the number of theory-based hypotheses plus a safeguard hypothesis (the complement or unconstrained)).
}
\examples{
  data(myGORICs)
  
}
\keyword{datasets}