\name{ZelazoKolb1972}
\alias{ZelazoKolb1972}
\docType{data}
\title{
Zelazo, Zelazo and Kolb Dataset - "Walking" in the newborn (4 treatment groups)
}
\description{
The Zelazo, Zelazo and Kolb (1972) dataset consists of ages (in months) at which 
an infant starts to walk alone from four different treatment groups (Active, 
Passive, Control, No). 
}

\usage{data(ZelazoKolb1972)}
\format{
  A data frame of 23 observations of 4 treatment variables.
  \describe{
    \item{\code{Group 1: Active-exercise}}{}
    \item{\code{Group 2: Passive-exercise}}{}
    \item{\code{Group 3: Control}}{8-Week control group}
    \item{\code{Group 4: No-exercise}}{}
  }
}
\references{Zelazo, P.R., Zelazo, N.A., and Kolb, S. (1972). "Walking in the 
Newborn". Science, New Series, 176, 314-315
}
\examples{
head(ZelazoKolb1972)
}
