\name{evSyn}
\alias{evSyn}
\alias{evSyn.default}
\alias{evSyn.est}
\alias{evSyn.ICweights}
\alias{evSyn.ICvalues}
\alias{evSyn.LL}
\alias{plot.evSyn}

\title{GORIC(A) Evidence synthesis}
\description{GORIC(A) evidence synthesis aggregates the evidence for theory-based 
hypotheses from multiple studies that may use diverse designs to investigate the 
same central theory.}

\usage{evSyn(object, ...)

\method{plot}{evSyn}(x, \dots)
}


\arguments{
  \item{object}{an object}
  \item{x}{an object of class \code{evSyn}}
  \item{\ldots}{This depends on the class of the object.}
}

\details{details
}

\value{output
}

\references{
 refs
}

\author{Leonard Vanbrabant and Rebecca Kuiper}

\examples{
## example
}


